+++ frontend/components/applicant/applicantNode/CustomResource.component.tsx
@@ -1,24 +1,13 @@
 import Txt from "@/components/common/Txt.component";
 import { ApplicantReq } from "@/src/apis/applicant";
 import { applicantDataFinder } from "@/src/functions/finder";
-import Link from "next/link";
-
+import Portfolio from "./Portfolio.component";
 interface ApplicantResourceProps {
   data: ApplicantReq[];
   postId: string;
 }
 
 const ApplicantResource = ({ data, postId }: ApplicantResourceProps) => {
-  const regex = /[\s,;|]+/;
-
-  const portfolio = applicantDataFinder(data, "portfolio")
-    .split(regex)
-    .map((url: string) => url.trim());
-
-  const file = applicantDataFinder(data, "fileUrl")
-    .split(regex)
-    .map((url: string) => url.trim());
-
   return (
     <>
       <div className="flex flex-col gap-1 mb-2">
@@ -49,29 +38,7 @@ const ApplicantResource = ({ data, postId }: ApplicantResourceProps) => {
         </div>
       </div>
       <div className="flex flex-col gap-4">
-        <Txt typography="h4">포트폴리오</Txt>
-        <div className="flex gap-4">
-          <div className="flex-1 flex flex-col">
-            <Txt typography="h6">링크</Txt>
-            {portfolio.map((url: string, index: number) => {
-              return (
-                <Link href={url} target="_blank" key={index}>
-                  <Txt className="break-all">{url}</Txt>
-                </Link>
-              );
-            })}
-          </div>
-          <div className="flex-1 flex flex-col">
-            <Txt typography="h6">파일</Txt>
-            {file.map((url: string, index: number) => {
-              return (
-                <Link href={url} target="_blank" key={index}>
-                  <Txt className="break-all">{url}</Txt>
-                </Link>
-              );
-            })}
-          </div>
-        </div>
+        <Portfolio data={data} />
       </div>
     </>
   );
